fix(core data): remove providers, move their logic to stores, rework updating in the filter, change reloadData to insert/delete sectionss
