+++ Tracker/CoreData/Stores/DataStoreManager.swift
@@ -24,9 +24,8 @@ final class DataStoreManager {
     lazy var persistentContainer: NSPersistentContainer = {
         let container = NSPersistentContainer(name: "Tracker")
         container.loadPersistentStores(completionHandler: { (storeDescription, error) in
-            print(storeDescription.url!)
             if let error = error as? NSError {
-                fatalError("Unresolved error \(error), \(error.userInfo)")
+                assertionFailure("Unresolved error \(error), \(error.userInfo)")
             }
             
         })
