+++ Tracker/CoreData/Stores/TrackerStore.swift
@@ -5,33 +5,162 @@
 //  Created by Danil Kazakov on 13.04.2025.
 //
 
-import Foundation
+import UIKit
 import CoreData
 
+struct TrackerStoreUpdate {
+    let insertedIndexPaths: [IndexPath]
+    let deletedIndexPaths: [IndexPath]
+    let insertedSections: IndexSet
+    let deletedSections: IndexSet
+}
+
+protocol TrackerStoreDelegate: AnyObject {
+    func didUpdateTrackers(_ update: TrackerStoreUpdate)
+}
 
 protocol TrackerStoreProtocol {
-    func addNewTracker(_ tracker: Tracker, categoryEntity: TrackerCategoryCoreData) throws
-    func getTrackerById(by id: UUID) -> TrackerCoreData?
+    var numberOfSections: Int { get }
+    var trackers: [Tracker] { get }
+    func nameOfSection(_ section: Int) -> String
+    func updateFilter(currentDate: Date, searchText: String?)
+    func numbersOfRowsInSections(in section: Int) -> Int
+    func tracker(at indexPath: IndexPath) -> Tracker?
+    func tracker(by id: UUID) -> TrackerCoreData?
+    func addNewTracker(tracker: Tracker, category: TrackerCategory) throws
 }
 
-final class TrackerStore {
-    private var context: NSManagedObjectContext
+final class TrackerStore: NSObject {
+    weak var delegate: TrackerStoreDelegate?
+    
+    private let context: NSManagedObjectContext
     private let categoryStore: TrackerCategoryStoreProtocol
     
-    init(context: NSManagedObjectContext) {
-        self.context = context
-        self.categoryStore = TrackerCategoryStore(context: context)
-    }
+    private var insertedIndexPaths: [IndexPath] = []
+    private var deletedIndexPaths: [IndexPath] = []
+    private var insertedSections: IndexSet = []
+    private var deletedSections: IndexSet = []
     
-    convenience init() {
-        self.init(context: DataStoreManager.shared.viewContext)
+    private lazy var fetchedResultsController: NSFetchedResultsController<TrackerCoreData> = {
+        let fetchRequest = TrackerCoreData.fetchRequest()
+        fetchRequest.sortDescriptors = [NSSortDescriptor(key: "createdAt", ascending: false)]
+        
+        let controller = NSFetchedResultsController(
+            fetchRequest: fetchRequest,
+            managedObjectContext: context,
+            sectionNameKeyPath: "category.name",
+            cacheName: nil
+        )
+        controller.delegate = self
+        try? controller.performFetch()
+        return controller
+    }()
+    
+    init(
+        context: NSManagedObjectContext = DataStoreManager.shared.viewContext,
+        categoryStore: TrackerCategoryStoreProtocol,
+        delegate: TrackerStoreDelegate
+    ) {
+        self.context = context
+        self.categoryStore = categoryStore
+        self.delegate = delegate
     }
 }
 
+// MARK: - Private Methods
+private extension TrackerStore {
+    func modelFrom(_ object: TrackerCoreData) -> Tracker {
+        Tracker(
+            id: object.id ?? UUID(),
+            name: object.name ?? "",
+            color: object.color as? UIColor ?? .clear,
+            emoji: object.emoji ?? "❓",
+            schedule: object.schedule.toWeekSet()
+        )
+    }
+}
 
 // MARK: - TrackerStoreProtocol
 extension TrackerStore: TrackerStoreProtocol {
-    func addNewTracker(_ tracker: Tracker, categoryEntity: TrackerCategoryCoreData) throws {
+    var numberOfSections: Int {
+        fetchedResultsController.sections?.count ?? 0
+    }
+    
+    var trackers: [Tracker] {
+        fetchedResultsController.fetchedObjects?.map { modelFrom($0) } ?? []
+    }
+    
+    func nameOfSection(_ section: Int) -> String {
+        fetchedResultsController.sections?[section].name ?? ""
+    }
+    
+    func numbersOfRowsInSections(in section: Int) -> Int {
+        fetchedResultsController.sections?[section].numberOfObjects ?? 0
+    }
+    
+    func tracker(at indexPath: IndexPath) -> Tracker? {
+        modelFrom(fetchedResultsController.object(at: indexPath))
+    }
+    
+    func tracker(by id: UUID) -> TrackerCoreData? {
+        fetchedResultsController.fetchedObjects?.first { $0.id == id }
+    }
+    
+    
+    func updateFilter(currentDate: Date, searchText: String?) {
+        let previousSectionsCount = fetchedResultsController.sections?.count ?? 0
+        let deletedSections = previousSectionsCount > 0 ? IndexSet(integersIn: 0..<previousSectionsCount) : IndexSet()
+        
+        var predicates: [NSPredicate] = []
+        
+        if let searchText, !searchText.isEmpty {
+            predicates.append(NSPredicate(format: "name CONTAINS[cd] %@", searchText))
+        }
+        
+        let calendar = Calendar.current
+        let weekday = calendar.component(.weekday, from: currentDate)
+        let weekdayStr = "\(weekday)"
+        let startOfDay = calendar.startOfDay(for: currentDate)
+        
+        // Предикат отображает трекеры по дням недели из их расписания.
+        // И если у трекера schedule == nil (нерегулярное событие), тогда:
+        // 1. Он не выполнен ни разу - отображаем каждый день
+        // 2. Если был выполнен хотя бы раз - отображаем только в день выполнения.
+        let predicate = NSPredicate(
+            format: """
+            (schedule != nil AND schedule CONTAINS[c] %@) OR
+            (schedule == nil AND (
+                SUBQUERY(records, $r, $r.tracker == SELF).@count == 0 OR
+                SUBQUERY(records, $r, $r.date == %@).@count > 0
+            ))
+            """,
+            weekdayStr, startOfDay as NSDate
+        )
+        
+        predicates.append(predicate)
+        
+        fetchedResultsController.fetchRequest.predicate = NSCompoundPredicate(andPredicateWithSubpredicates: predicates)
+        
+        do {
+            try fetchedResultsController.performFetch()
+            let newSectionsCount = fetchedResultsController.sections?.count ?? 0
+            let insertedSections = newSectionsCount > 0 ? IndexSet(integersIn: 0..<newSectionsCount) : IndexSet()
+            
+            // Отправляем делегату разницу между секциями до фетча и после
+            let update = TrackerStoreUpdate(
+                insertedIndexPaths: [],
+                deletedIndexPaths: [],
+                insertedSections: insertedSections,
+                deletedSections: deletedSections
+            )
+            delegate?.didUpdateTrackers(update)
+        } catch {
+            print("Failed to fetch trackers: \(error)")
+        }
+    }
+    
+    func addNewTracker(tracker: Tracker, category: TrackerCategory) throws {
+        let categoryEntity = try categoryStore.fetchOrCreateCategory(from: category)
         let trackerEntity = TrackerCoreData(context: context)
         trackerEntity.id = tracker.id
         trackerEntity.name = tracker.name
@@ -42,17 +171,60 @@ extension TrackerStore: TrackerStoreProtocol {
         trackerEntity.category = categoryEntity
         DataStoreManager.shared.saveContext()
     }
+}
+
+// MARK: - NSFetchedResultsControllerDelegate
+extension TrackerStore: NSFetchedResultsControllerDelegate {
+    func controllerWillChangeContent(_ controller: NSFetchedResultsController<any NSFetchRequestResult>) {
+        insertedIndexPaths = []
+        deletedIndexPaths = []
+        insertedSections = []
+        deletedSections = []
+    }
     
-    func getTrackerById(by id: UUID) -> TrackerCoreData? {
-        let fetchRequest = TrackerCoreData.fetchRequest()
-        fetchRequest.predicate = NSPredicate(format: "id == %@", id as CVarArg)
-        
-        do {
-            let results = try context.fetch(fetchRequest)
-            return results.first
-        } catch {
-            print("Failed to fetch tracker: \(error)")
-            return nil
+    func controllerDidChangeContent(_ controller: NSFetchedResultsController<any NSFetchRequestResult>) {
+        delegate?.didUpdateTrackers(TrackerStoreUpdate(
+            insertedIndexPaths: insertedIndexPaths,
+            deletedIndexPaths: deletedIndexPaths,
+            insertedSections: insertedSections,
+            deletedSections: deletedSections
+        ))
+        insertedIndexPaths = []
+        deletedIndexPaths = []
+        insertedSections = []
+        deletedSections = []
+    }
+    
+    func controller(
+        _ controller: NSFetchedResultsController<NSFetchRequestResult>,
+        didChange sectionInfo: NSFetchedResultsSectionInfo,
+        atSectionIndex sectionIndex: Int,
+        for type: NSFetchedResultsChangeType
+    ) {
+        switch type {
+        case .insert:
+            insertedSections.insert(sectionIndex)
+        case .delete:
+            deletedSections.insert(sectionIndex)
+        default:
+            break
+        }
+    }
+    
+    func controller(
+        _ controller: NSFetchedResultsController<NSFetchRequestResult>,
+        didChange anObject: Any,
+        at indexPath: IndexPath?,
+        for type: NSFetchedResultsChangeType,
+        newIndexPath: IndexPath?
+    ) {
+        switch type {
+        case .insert:
+            if let newIndexPath { insertedIndexPaths.append(newIndexPath) }
+        case .delete:
+            if let indexPath { deletedIndexPaths.append(indexPath) }
+        default:
+            break
         }
     }
 }
